@@ -177,6 +177,7 @@ class StudentsController < ApplicationController
   end
 
   def create
+    flash = {}
     @student = Student.new(params[:student])
 
     if @student.save
@@ -213,6 +214,7 @@ class StudentsController < ApplicationController
   end
 
   def update 
+    flash = {}
     @student = Student.find(params[:id])
 
     if @student.update_attributes(params[:student])
@@ -253,6 +255,7 @@ class StudentsController < ApplicationController
   end
 
   def upload_image
+    flash = {}
     @student = Student.find(params[:id])
     if @student.update_attributes(params[:student])
       flash[:notice] = "Imagen actualizada."
@@ -270,8 +273,9 @@ class StudentsController < ApplicationController
   end
 
   def upload_file
+    flash = {}
     params[:student_file]['file'].each do |f|
-      @student_file = StudentFile.new(f)
+      @student_file = StudentFile.new
       @student_file.student_id = params[:student_file]['student_id']
       @student_file.file = f
       @student_file.description = f.original_filename
@@ -301,6 +305,7 @@ class StudentsController < ApplicationController
   end
 
   def create_advance
+    flash = {}
     @student = Student.find(params[:student_id])
     if @student.update_attributes(params[:student])
       flash[:notice] = "Nuevo avance creado."
@@ -311,6 +316,7 @@ class StudentsController < ApplicationController
   end
 
   def assign_thesis_number
+    flash = {}
     @student = Student.find(params[:student_id])
     @student.thesis.set_number
     if @student.save
@@ -402,7 +408,8 @@ class StudentsController < ApplicationController
         @is_pdf = true
         html = render_to_string(:layout => false , :action => "schedule_table.html.haml")
         kit = PDFKit.new(html, :page_size => 'Letter')
-        kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/pdf.css"
+        #kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/pdf.css"
+        kit.stylesheets << "http://posgrado.cimav.edu.mx" + view_context.asset_path('pdf.css')
         filename = "horario-#{@ts.student_id}-#{@ts.term_id}.pdf"
         send_data(kit.to_pdf, :filename => filename, :type => 'application/pdf')
         return # to avoid double render call
@@ -425,7 +432,8 @@ class StudentsController < ApplicationController
         @is_pdf = true
         html = render_to_string(:layout => false , :action => "term_grades.html.haml")
         kit = PDFKit.new(html, :page_size => 'Letter')
-        kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/pdf.css"
+        #kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/pdf.css"
+        kit.stylesheets << "http://posgrado.cimav.edu.mx" + view_context.asset_path('pdf.css')
         filename = "boleta-#{@ts.student_id}-#{@ts.term_id}.pdf"
         send_data(kit.to_pdf, :filename => filename, :type => 'application/pdf')
         return # to avoid double render call
@@ -459,7 +467,8 @@ class StudentsController < ApplicationController
         @is_pdf = true
         html = render_to_string(:layout => false , :action => "id_card.html.haml")
         kit = PDFKit.new(html, :page_size => 'Legal', :orientation => 'Landscape', :margin_top    => '0',:margin_right  => '0', :margin_bottom => '0', :margin_left   => '0')
-        kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/card.css"
+        # kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/card.css"
+        kit.stylesheets << "http://posgrado.cimav.edu.mx" + view_context.asset_path('card.css')
         filename = "ID-#{@student.card}.pdf"
         send_data(kit.to_pdf, :filename => filename, :type => 'application/pdf')
         return # to avoid double render call
@@ -480,7 +489,9 @@ class StudentsController < ApplicationController
         @is_pdf = true
         html = render_to_string(:layout => false , :action => "_kardex.html.haml")
         kit = PDFKit.new(html, :page_size => 'Letter')
-        kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/pdf.css"
+        # kit.stylesheets << "#{Rails.root}/public/stylesheets/compiled/pdf.css"
+        kit.stylesheets << "http://posgrado.cimav.edu.mx" + view_context.asset_path('pdf.css')
+
         filename = "kardex-#{@student.id}.pdf"
         send_data(kit.to_pdf, :filename => filename, :type => 'application/pdf')
         return # to avoid double render call
