@@ -13,12 +13,12 @@ class StudentImageUploader < CarrierWave::Uploader::Base
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
   def store_dir
-    "images/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
+    "/images/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
   end
 
   # Provide a default URL as a default if there hasn't been a file uploaded:
   def default_url
-    "/images/#{model.class.to_s.underscore}/fallback/" + [version_name, "default.jpg"].compact.join('_')
+    "/assets/#{model.class.to_s.underscore}/fallback/" + [version_name, "default.jpg"].compact.join('_')
   end
 
   # Process files as they are uploaded:
@@ -38,7 +38,7 @@ class StudentImageUploader < CarrierWave::Uploader::Base
   end
 
   # Add a white list of extensions which are allowed to be uploaded.
-  # For images you might use something like this:
+  # For assets you might use something like this:
   # def extension_white_list
   #   %w(jpg jpeg gif png)
   # end
