@@ -3,5 +3,5 @@
 # Add new mime types for use in respond_to blocks:
 # Mime::Type.register "text/richtext", :rtf
 # Mime::Type.register_alias "text/html", :iphone
+#
 Mime::Type.register "application/vnd.ms-excel", :xls
-Mime::Type.register_alias "application/pdf", :pdf
